backend/requirements.txt
fastapi
uvicorn
numpy
onnxruntime
Pillow
python-multipart